+++ src/assets/icons/index.tsx
@@ -173,55 +173,3 @@ export const GoogleStore = () => (
     </defs>
   </svg>
 );
-
-export const ConcentricCirclesIcon = () => (
-  <svg width='24' height='24' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'>
-    <path
-      fill-rule='evenodd'
-      clip-rule='evenodd'
-      d='M12 21C16.9705 21 21 16.9705 21 12C21 7.02944 16.9705 3 12 3C7.02944 3 3 7.02944 3 12C3 16.9705 7.02944 21 12 21ZM12 19.5C16.1421 19.5 19.5 16.1421 19.5 12C19.5 7.85787 16.1421 4.5 12 4.5C7.85787 4.5 4.5 7.85787 4.5 12C4.5 16.1421 7.85787 19.5 12 19.5Z'
-      fill='#06AED5'
-    />
-    <path
-      fill-rule='evenodd'
-      clip-rule='evenodd'
-      d='M12 18C15.3137 18 18 15.3137 18 12C18 8.68629 15.3137 6 12 6C8.68629 6 6 8.68629 6 12C6 15.3137 8.68629 18 12 18ZM12 16.5C14.4853 16.5 16.5 14.4853 16.5 12C16.5 9.51473 14.4853 7.5 12 7.5C9.51473 7.5 7.5 9.51473 7.5 12C7.5 14.4853 9.51473 16.5 12 16.5Z'
-      fill='#06AED5'
-    />
-    <path
-      d='M12 15C13.6569 15 15 13.6569 15 12C15 10.3431 13.6569 9 12 9C10.3431 9 9 10.3431 9 12C9 13.6569 10.3431 15 12 15Z'
-      fill='#06AED5'
-    />
-  </svg>
-);
-
-export const HeartIcon = () => (
-  <svg width='22' height='22' viewBox='0 0 22 22' fill='none' xmlns='http://www.w3.org/2000/svg'>
-    <path
-      d='M6.5 0.5C3.18629 0.5 0.5 3.19723 0.5 6.52442C0.5 12.0812 6.21721 18.048 9.21418 20.7941C10.2374 21.7318 11.7626 21.7318 12.7858 20.7941C15.7828 18.048 21.5 12.0812 21.5 6.52442C21.5 3.19723 18.8137 0.5 15.5 0.5C13.708 0.5 12.0994 1.28883 11 2.53954C9.90059 1.28883 8.29204 0.5 6.5 0.5Z'
-      fill='#FF540B'
-    />
-  </svg>
-);
-
-export const PersonIcon = () => (
-  <svg width='24' height='24' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'>
-    <path
-      fill-rule='evenodd'
-      clip-rule='evenodd'
-      d='M12 22.5C14.7848 22.5 17.4555 21.3938 19.4246 19.4246C21.3938 17.4555 22.5 14.7848 22.5 12C22.5 9.21523 21.3938 6.54451 19.4246 4.57538C17.4555 2.60625 14.7848 1.5 12 1.5C9.21523 1.5 6.54451 2.60625 4.57538 4.57538C2.60625 6.54451 1.5 9.21523 1.5 12C1.5 14.7848 2.60625 17.4555 4.57538 19.4246C6.54451 21.3938 9.21523 22.5 12 22.5ZM13.5 6C13.5 6.39782 13.342 6.77936 13.0607 7.06066C12.7794 7.34196 12.3978 7.5 12 7.5C11.6022 7.5 11.2206 7.34196 10.9393 7.06066C10.658 6.77936 10.5 6.39782 10.5 6C10.5 5.60218 10.658 5.22064 10.9393 4.93934C11.2206 4.65804 11.6022 4.5 12 4.5C12.3978 4.5 12.7794 4.65804 13.0607 4.93934C13.342 5.22064 13.5 5.60218 13.5 6ZM6.825 9C6.6219 9.00225 6.42676 9.07935 6.27697 9.21653C6.12718 9.35371 6.03325 9.54133 6.01319 9.74346C5.99313 9.94558 6.04833 10.148 6.16823 10.312C6.28813 10.4759 6.4643 10.5899 6.663 10.632L9.591 11.217L9.0585 18.672C9.03976 18.8741 9.10128 19.0754 9.22978 19.2325C9.35827 19.3895 9.54342 19.4897 9.74519 19.5114C9.94696 19.5331 10.1492 19.4745 10.3081 19.3483C10.467 19.2221 10.5699 19.0384 10.5945 18.837L10.935 16.4505C11.0355 15.756 11.2995 15 12 15C12.7005 15 12.9645 15.756 13.065 16.4505L13.4055 18.8355C13.431 19.036 13.5341 19.2186 13.6926 19.344C13.8512 19.4694 14.0526 19.5276 14.2536 19.5062C14.4546 19.4848 14.6392 19.3855 14.7678 19.2295C14.8964 19.0736 14.9588 18.8734 14.9415 18.672L14.409 11.217L17.337 10.632C17.5356 10.5899 17.7117 10.4761 17.8316 10.3123C17.9515 10.1485 18.0068 9.94621 17.987 9.74419C17.9671 9.54216 17.8734 9.35455 17.7239 9.21725C17.5744 9.07995 17.3795 9.0026 17.1765 9H6.825Z'
-      fill='#AC54F0'
-    />
-  </svg>
-);
-
-export const StarIcon = () => (
-  <svg width='24' height='24' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'>
-    <path
-      fill-rule='evenodd'
-      clip-rule='evenodd'
-      d='M22.4162 9.07884C22.3197 8.82238 22.1469 8.60116 21.9204 8.4447C21.6939 8.28823 21.4248 8.20397 21.1488 8.20316H15.6625C15.5693 8.20322 15.4781 8.17454 15.4019 8.12108C15.3257 8.06762 15.2681 7.99202 15.2372 7.9047L13.2677 2.37092C13.1712 2.11515 12.9983 1.8947 12.772 1.73904C12.5457 1.5834 12.2768 1.5 12.0014 1.5C11.7259 1.5 11.4571 1.5834 11.2308 1.73904C11.0044 1.8947 10.8315 2.11515 10.735 2.37092L10.7305 2.38521L8.76562 7.9047C8.73469 7.99187 8.67724 8.06737 8.60122 8.12082C8.52519 8.17426 8.43432 8.20303 8.34113 8.20316H2.853C2.57523 8.20295 2.30412 8.28756 2.07659 8.4455C1.84904 8.60343 1.67609 8.82702 1.58124 9.08586C1.48641 9.34469 1.47428 9.62619 1.54653 9.89209C1.61879 10.158 1.77189 10.3954 1.98503 10.572L6.65745 14.4142C6.72735 14.4718 6.77808 14.5488 6.80301 14.6354C6.82794 14.722 6.8259 14.814 6.79714 14.8995L4.83408 20.7352C4.74336 21.0051 4.74158 21.2965 4.82898 21.5674C4.91637 21.8384 5.08844 22.0746 5.32026 22.2421C5.55209 22.4097 5.83166 22.5 6.11855 22.4994C6.40544 22.4991 6.68478 22.4082 6.91614 22.2401L11.7346 18.7372C11.8118 18.6812 11.9051 18.6509 12.0009 18.6509C12.0967 18.6509 12.19 18.6812 12.2673 18.7372L17.084 22.2392C17.3151 22.4079 17.5945 22.4993 17.8818 22.5C18.1689 22.5007 18.4488 22.4109 18.681 22.2434C18.9131 22.0758 19.0854 21.8393 19.173 21.5682C19.2604 21.297 19.2586 21.0052 19.1677 20.7352L17.2047 14.8959C17.1759 14.8105 17.174 14.7184 17.1989 14.6319C17.2238 14.5453 17.2744 14.4682 17.3443 14.4107L22.0259 10.563C22.2363 10.3849 22.3864 10.1472 22.4562 9.88175C22.5261 9.61633 22.5121 9.33609 22.4162 9.07884Z'
-      fill='#FFA500'
-    />
-  </svg>
-);
