+++ src/index.css
@@ -32,7 +32,7 @@
   }
 
   .dark {
-    --background: 130, 16%, 7%, 1;
+    --background: 222.2 84% 4.9%;
     --foreground: 210 40% 98%;
     --card: 222.2 84% 4.9%;
     --card-foreground: 210 40% 98%;
