+++ src/main.tsx
@@ -1,20 +1,14 @@
-import './App.css';
-import './index.css';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
+import App from './App.tsx';
+import './index.css';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import ErrorPage from './error-page.tsx';
-import Home from './pages/home.tsx';
-import About from './pages/about.tsx';
 const router = createBrowserRouter([
   {
     path: '/',
-    element: <Home />,
+    element: <App />,
     errorElement: <ErrorPage />
-  },
-  {
-    path: '/about',
-    element: <About />
   }
 ]);
 createRoot(document.getElementById('root')!).render(
